refactor: code review fixes
