+++ packages/polaris-common/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@enigmatis/polaris-common",
-  "version": "1.10.1-snapshot.0",
+  "version": "1.10.1",
   "description": "common types for polaris",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
