+++ packages/polaris-common/src/utils/common-methods.ts
