+++ packages/polaris-core/guides/online-paging-without-relay/README.md
@@ -31,8 +31,12 @@ Note that you have to use the `OnlinePaginatedResolver` as your return type, and
 
 In order to use the online paging correctly you'll need to use the compatible request headers.
 - `page-size` - *optional* if set to a number, it will be the page size. If not, the size will be taken from the default configurations.
-- `data-version` - *optional* if not set, will always fetch the first page according to the page-size. Serves as the cursor of the first entity that will be fetched in the page.
-- `last-id-in-dv` - *optional* if not set will always fetch the first page according to the page-size and the data-version header. Serves as the cursor of the last id of the entity that was fetched in the previous page.
+- `data-version` - *optional* if not set, will always fetch the first page according to the page-size. Serves as the primary cursor.
+- `last-id-in-dv` - *optional* serves as the secondary cursor.
+
+
+**Pay Attention!** The entities are sorted firstly by the `data-version` header and secondly by the`last-id-in-dv` header.
+The cursor runs according to that sort, so only sending the `last-id-in-dv` header won't serve the purpose of the cursor since the data version of the entity could've been changed.
 
 For example:
 ```json
