+++ packages/polaris-core/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@enigmatis/polaris-core",
-    "version": "2.0.1-snapshot.0",
+    "version": "2.0.1",
     "description": "Create a node js graphql service easily",
     "main": "dist/src/index.js",
     "types": "dist/src/index.d.ts",
@@ -30,12 +30,12 @@
     },
     "homepage": "https://github.com/Enigmatis/polaris-united/tree/master/packages/polaris-core#readme",
     "dependencies": {
-        "@enigmatis/polaris-common": "^1.10.1-snapshot.0",
-        "@enigmatis/polaris-graphql-logger": "^2.3.3-snapshot.0",
+        "@enigmatis/polaris-common": "^1.10.1",
+        "@enigmatis/polaris-graphql-logger": "^2.3.3",
         "@enigmatis/polaris-logs": "5.0.0-beta.2",
-        "@enigmatis/polaris-middlewares": "^1.8.1-snapshot.0",
-        "@enigmatis/polaris-schema": "^1.5.3-snapshot.0",
-        "@enigmatis/polaris-typeorm": "^2.7.1-snapshot.0",
+        "@enigmatis/polaris-middlewares": "^1.8.1",
+        "@enigmatis/polaris-schema": "^1.5.3",
+        "@enigmatis/polaris-typeorm": "^2.7.1",
         "apollo-server-express": "^2.17.0",
         "apollo-server-plugin-base": "0.6.10",
         "express": "^4.17.1",
