+++ packages/polaris-graphql-logger/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@enigmatis/polaris-graphql-logger",
-    "version": "2.3.3-snapshot.0",
+    "version": "2.3.3",
     "description": "A library that organizes graphql logs",
     "main": "dist/src/main.js",
     "types": "dist/src/main.d.ts",
@@ -28,7 +28,7 @@
     },
     "homepage": "https://github.com/Enigmatis/polaris-graphql-logger#readme",
     "dependencies": {
-        "@enigmatis/polaris-common": "^1.10.1-snapshot.0",
+        "@enigmatis/polaris-common": "^1.10.1",
         "@enigmatis/polaris-logs": "5.0.0-beta.2",
         "clean-deep": "^3.3.0",
         "winston": "^3.2.1"
