+++ packages/polaris-middlewares/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@enigmatis/polaris-middlewares",
-    "version": "1.8.1-snapshot.0",
+    "version": "1.8.1",
     "description": "handling middlewares in polaris",
     "main": "dist/src/index.js",
     "types": "dist/src/index.d.ts",
@@ -32,10 +32,10 @@
     },
     "homepage": "https://github.com/Enigmatis/polaris-united/tree/master/packages/polaris-middlewares#readme",
     "dependencies": {
-        "@enigmatis/polaris-common": "^1.10.1-snapshot.0",
-        "@enigmatis/polaris-graphql-logger": "^2.3.3-snapshot.0",
-        "@enigmatis/polaris-schema": "^1.5.3-snapshot.0",
-        "@enigmatis/polaris-typeorm": "^2.7.1-snapshot.0",
+        "@enigmatis/polaris-common": "^1.10.1",
+        "@enigmatis/polaris-graphql-logger": "^2.3.3",
+        "@enigmatis/polaris-schema": "^1.5.3",
+        "@enigmatis/polaris-typeorm": "^2.7.1",
         "graphql-extensions": "^0.10.3",
         "uuid": "^8.3.1"
     },
