+++ packages/polaris-permissions/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@enigmatis/polaris-permissions",
-  "version": "1.2.7-snapshot.0",
+  "version": "1.2.7",
   "description": "Permissions service wrapper for polaris framework",
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
@@ -33,7 +33,7 @@
   },
   "homepage": "https://github.com/Enigmatis/polaris-united/tree/master/packages/polaris-permissions#readme",
   "dependencies": {
-    "@enigmatis/polaris-common": "^1.10.1-snapshot.0",
+    "@enigmatis/polaris-common": "^1.10.1",
     "axios": "^0.20.0",
     "tunnel": "0.0.6"
   },
