+++ packages/polaris-test/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@enigmatis/polaris-test",
-  "version": "1.4.1-snapshot.0",
+  "version": "1.4.1",
   "description": "test repository",
   "license": "ISC",
   "main": "dist/index.js",
@@ -27,8 +27,8 @@
     "typescript"
   ],
   "dependencies": {
-    "@enigmatis/polaris-core": "^2.0.1-snapshot.0",
-    "@enigmatis/polaris-nest": "^1.4.3-snapshot.0",
+    "@enigmatis/polaris-core": "^2.0.1",
+    "@enigmatis/polaris-nest": "^1.4.3",
     "@nestjs/common": "7.0.5",
     "@nestjs/core": "7.0.5",
     "@nestjs/graphql": "^7.3.9",
