+++ packages/polaris-test/tests/date-filter.test.ts
@@ -11,7 +11,7 @@ function sleep(ms: number) {
     return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-describe.skip('date filter tests', () => {
+describe('date filter tests', () => {
     describe('creation time tests', () => {
         test.each(createServers())(
             'fetch books by date filter, books fetched successfully',
@@ -190,7 +190,8 @@ describe.skip('date filter tests', () => {
             },
         );
     });
-    describe('multiple queries with filter dates', () => {
+    // TODO: find solution for multiple date filter queries
+    describe.skip('multiple queries with filter dates', () => {
         test.each(createServersWithoutNestServer())(
             'execute multiple queries in the same request, execution executed successfully for each query',
             async (server) => {
