+++ packages/polaris-typeorm/src/utils/query-builder-util.ts
@@ -4,24 +4,24 @@ import { SelectQueryBuilder } from 'typeorm';
 const addDateRangeFilterByFieldName = (
     qb: SelectQueryBuilder<any>,
     dateRangeFilter: DateRangeFilter,
-    entityTableName: string,
+    alias: string,
     filteredFieldName: string,
 ) => {
     if (dateRangeFilter?.gte) {
-        qb.andWhere(`${entityTableName}.${filteredFieldName} > :gte`, {
+        qb.andWhere(`${alias}.${filteredFieldName} > :gte`, {
             gte: dateRangeFilter?.gte,
         });
     } else if (dateRangeFilter?.gt) {
-        qb.andWhere(`${entityTableName}.${filteredFieldName} >= :gt`, {
+        qb.andWhere(`${alias}.${filteredFieldName} >= :gt`, {
             gt: dateRangeFilter?.gt,
         });
     }
     if (dateRangeFilter?.lte) {
-        qb.andWhere(`${entityTableName}.${filteredFieldName} < :lte`, {
+        qb.andWhere(`${alias}.${filteredFieldName} < :lte`, {
             lte: dateRangeFilter?.lte,
         });
     } else if (dateRangeFilter?.lt) {
-        qb.andWhere(`${entityTableName}.${filteredFieldName} <= :lt`, {
+        qb.andWhere(`${alias}.${filteredFieldName} <= :lt`, {
             lt: dateRangeFilter?.lt,
         });
     }
@@ -30,13 +30,13 @@ const addDateRangeFilterByFieldName = (
 export const addDateRangeCriteria = (
     qb: SelectQueryBuilder<any>,
     dateRangeFilter: EntityFilter,
-    entityTableName: string,
+    alias: string,
 ) => {
     if (dateRangeFilter.creationTimeFilter) {
         addDateRangeFilterByFieldName(
             qb,
             dateRangeFilter.creationTimeFilter,
-            entityTableName,
+            alias,
             'creationTime',
         );
     }
@@ -44,7 +44,7 @@ export const addDateRangeCriteria = (
         addDateRangeFilterByFieldName(
             qb,
             dateRangeFilter.lastUpdateTimeFilter,
-            entityTableName,
+            alias,
             'lastUpdateTime',
         );
     }
